Changed the file structure of the project.
