+++ .src/Tests/BoardTests.java
@@ -2,25 +2,13 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class Tests
+public class BoardTests
 {
   private Board board;
-  private Game game;
   @BeforeEach
   public void setUp()
   {
-    game = new Game();
     board = new Board();
-  }
-
-
-  @Test
-  public void testGameInit()
-  {
-    Assertions.assertEquals(0, game.getTurn());
-    Assertions.assertNotNull(game.getBoard());
-    Assertions.assertNotNull(game.getHuman());
-    Assertions.assertNotNull(game.getBot());
   }
 
   @Test
